Moved the GsimLogicTree class in hazardlib
